returned byte[] interface back, Span<T> is also avaible via BlingFireUtils2 static class
