+++ nuget/test/BlingUtilsTest.csproj
@@ -12,7 +12,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="BlingFireNuget" Version="0.1.5" />
+    <PackageReference Include="BlingFireNuget" Version="0.1.6.2" />
   </ItemGroup>
 
 </Project>
